store discount as big decimal not as integer

+++ delivery/src/main/java/com/storyous/delivery/common/api/model.kt
@@ -11,7 +11,7 @@ data class DeliveryOrder(
     val deliveryTime: Date,
     val deliveryOnTime: Boolean,
     @DeliveryType val deliveryType: String,
-    val discountWithVat: Int?,
+    val discountWithVat: BigDecimal?,
     val customer: Customer,
     val desk: Desk?,
     val items: List<DeliveryItem>,
